extract shared row rendering module
